feat: added the ability to generate a path based on the second FunctionId argument
