+++ internal/xsql/stmt.go
@@ -28,7 +28,7 @@ var (
 
 func (stmt *Stmt) QueryContext(ctx context.Context, args []driver.NamedValue) (_ driver.Rows, finalErr error) {
 	onDone := trace.DatabaseSQLOnStmtQuery(stmt.conn.connector.Trace(), &ctx,
-		stack.FunctionID("", stack.Package("database/sql")),
+		stack.FunctionID("database/sql.(*Stmt).QueryContext", stack.Package("database/sql")),
 		stmt.ctx, stmt.sql,
 	)
 	defer func() {
@@ -49,7 +49,7 @@ func (stmt *Stmt) QueryContext(ctx context.Context, args []driver.NamedValue) (_
 
 func (stmt *Stmt) ExecContext(ctx context.Context, args []driver.NamedValue) (_ driver.Result, finalErr error) {
 	onDone := trace.DatabaseSQLOnStmtExec(stmt.conn.connector.Trace(), &ctx,
-		stack.FunctionID("", stack.Package("database/sql")),
+		stack.FunctionID("database/sql.(*Stmt).ExecContext", stack.Package("database/sql")),
 		stmt.ctx, stmt.sql,
 	)
 	defer func() {
@@ -76,7 +76,7 @@ func (stmt *Stmt) Close() (finalErr error) {
 	var (
 		ctx    = stmt.ctx
 		onDone = trace.DatabaseSQLOnStmtClose(stmt.conn.connector.Trace(), &ctx,
-			stack.FunctionID("", stack.Package("database/sql")),
+			stack.FunctionID("database/sql.(*Stmt).Close", stack.Package("database/sql")),
 		)
 	)
 	defer func() {
