+++ internal/xsql/tx.go
@@ -39,7 +39,7 @@ func (tx *Tx) Commit() (finalErr error) {
 	var (
 		ctx    = tx.ctx
 		onDone = trace.DatabaseSQLOnTxCommit(tx.conn.connector.Trace(), &ctx,
-			stack.FunctionID("", stack.Package("database/sql")),
+			stack.FunctionID("database/sql.(*Tx).Commit", stack.Package("database/sql")),
 			tx,
 		)
 	)
@@ -62,7 +62,7 @@ func (tx *Tx) Rollback() (finalErr error) {
 	var (
 		ctx    = tx.ctx
 		onDone = trace.DatabaseSQLOnTxRollback(tx.conn.connector.Trace(), &ctx,
-			stack.FunctionID("", stack.Package("database/sql")),
+			stack.FunctionID("database/sql.(*Tx).Rollback", stack.Package("database/sql")),
 			tx,
 		)
 	)
@@ -82,7 +82,7 @@ func (tx *Tx) QueryContext(ctx context.Context, sql string, args []driver.NamedV
 	_ driver.Rows, finalErr error,
 ) {
 	onDone := trace.DatabaseSQLOnTxQuery(tx.conn.connector.Trace(), &ctx,
-		stack.FunctionID("", stack.Package("database/sql")),
+		stack.FunctionID("database/sql.(*Tx).QueryContext", stack.Package("database/sql")),
 		tx.ctx, tx, sql,
 	)
 	defer func() {
@@ -115,7 +115,7 @@ func (tx *Tx) ExecContext(ctx context.Context, sql string, args []driver.NamedVa
 	_ driver.Result, finalErr error,
 ) {
 	onDone := trace.DatabaseSQLOnTxExec(tx.conn.connector.Trace(), &ctx,
-		stack.FunctionID("", stack.Package("database/sql")),
+		stack.FunctionID("database/sql.(*Tx).ExecContext", stack.Package("database/sql")),
 		tx.ctx, tx, sql,
 	)
 	defer func() {
@@ -137,7 +137,7 @@ func (tx *Tx) ExecContext(ctx context.Context, sql string, args []driver.NamedVa
 
 func (tx *Tx) PrepareContext(ctx context.Context, sql string) (_ driver.Stmt, finalErr error) {
 	onDone := trace.DatabaseSQLOnTxPrepare(tx.conn.connector.Trace(), &ctx,
-		stack.FunctionID("", stack.Package("database/sql")),
+		stack.FunctionID("database/sql.(*Tx).PrepareContext", stack.Package("database/sql")),
 		tx.ctx, tx, sql,
 	)
 	defer func() {
